# compiler: optimise dir diff for output cleanup