--- a/compiler/package.yaml
+++ b/compiler/package.yaml
@@ -17,6 +17,7 @@ description:         Please see the README on GitHub at <https://github.com/paci
 dependencies:
 - base >= 4.7 && < 5
 - containers
+- data-ordlist
 - filepath
 - directory
 - text